Move argument util functions to manager and server
