+++ crates/modelardb_manager/src/main.rs
@@ -20,10 +20,9 @@ mod error;
 mod metadata;
 mod remote;
 
-use std::env;
 use std::sync::{Arc, LazyLock};
+use std::{env, process};
 
-use modelardb_common::arguments;
 use modelardb_storage::delta_lake::DeltaLake;
 use modelardb_types::flight::protocol;
 use tokio::runtime::Runtime;
@@ -109,11 +108,11 @@ fn main() -> Result<()> {
     // Create a Tokio runtime for executing asynchronous tasks.
     let runtime = Arc::new(Runtime::new()?);
 
-    let user_arguments = arguments::collect_command_line_arguments(3);
+    let user_arguments = collect_command_line_arguments(3);
     let user_arguments: Vec<&str> = user_arguments.iter().map(|arg| arg.as_str()).collect();
     let remote_data_folder_str = match user_arguments.as_slice() {
         &[remote_data_folder_str] => remote_data_folder_str,
-        _ => arguments::print_usage_and_exit_with_error("remote_data_folder"),
+        _ => print_usage_and_exit_with_error("remote_data_folder"),
     };
 
     let context = runtime.block_on(async {
@@ -146,4 +145,27 @@ fn main() -> Result<()> {
     })?;
 
     start_apache_arrow_flight_server(context, &runtime, *PORT)
+}
+
+/// Collect the command line arguments that this program was started with.
+pub fn collect_command_line_arguments(maximum_arguments: usize) -> Vec<String> {
+    let mut args = std::env::args();
+    args.next(); // Skip the executable.
+
+    // Collect at most the maximum number of command line arguments plus one. The plus one argument
+    // is collected to trigger the default pattern when parsing the command line arguments with
+    // pattern matching, making it possible to handle errors caused by too many arguments.
+    args.by_ref().take(maximum_arguments + 1).collect()
+}
+
+/// Prints a usage message with `parameters` appended to the name of the binary executing this
+/// function to stderr and exits with status code one to indicate that an error has occurred.
+pub fn print_usage_and_exit_with_error(parameters: &str) -> ! {
+    // The errors are consciously ignored as the program is terminating.
+    let binary_path = std::env::current_exe().unwrap();
+    let binary_name = binary_path.file_name().unwrap().to_str().unwrap();
+
+    // Punctuation at the end does not seem to be common in the usage message of Unix tools.
+    eprintln!("Usage: {binary_name} {parameters}");
+    process::exit(1);
 }
