+++ crates/modelardb_server/src/main.rs
@@ -25,10 +25,9 @@ mod manager;
 mod remote;
 mod storage;
 
-use std::env;
 use std::sync::{Arc, LazyLock};
+use std::{env, process};
 
-use modelardb_common::arguments::{self, collect_command_line_arguments};
 use tokio::runtime::Runtime;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
@@ -75,9 +74,7 @@ fn main() -> Result<()> {
     {
         cluster_mode_and_data_folders
     } else {
-        arguments::print_usage_and_exit_with_error(
-            "[server_mode] local_data_folder_url [manager_url]",
-        );
+        print_usage_and_exit_with_error("[server_mode] local_data_folder_url [manager_url]");
     };
 
     let context = Arc::new(runtime.block_on(Context::try_new(
@@ -111,6 +108,29 @@ fn main() -> Result<()> {
     remote::start_apache_arrow_flight_server(context, &runtime, *PORT)?;
 
     Ok(())
+}
+
+/// Collect the command line arguments that this program was started with.
+pub fn collect_command_line_arguments(maximum_arguments: usize) -> Vec<String> {
+    let mut args = std::env::args();
+    args.next(); // Skip the executable.
+
+    // Collect at most the maximum number of command line arguments plus one. The plus one argument
+    // is collected to trigger the default pattern when parsing the command line arguments with
+    // pattern matching, making it possible to handle errors caused by too many arguments.
+    args.by_ref().take(maximum_arguments + 1).collect()
+}
+
+/// Prints a usage message with `parameters` appended to the name of the binary executing this
+/// function to stderr and exits with status code one to indicate that an error has occurred.
+pub fn print_usage_and_exit_with_error(parameters: &str) -> ! {
+    // The errors are consciously ignored as the program is terminating.
+    let binary_path = std::env::current_exe().unwrap();
+    let binary_name = binary_path.file_name().unwrap().to_str().unwrap();
+
+    // Punctuation at the end does not seem to be common in the usage message of Unix tools.
+    eprintln!("Usage: {binary_name} {parameters}");
+    process::exit(1);
 }
 
 /// Register a handler to execute when CTRL+C is pressed. The handler takes an exclusive lock for
